remove unsued last_song code.
previously, i planned for both music override and playlist songs to be resumed from their last position, but decided against it because
it was tricky to managed.

for example, if you launch the eshop and it starts playing a song, then you close the eshop, then reopen it, tune would resume from the last
position.